… UPGRADE_LOG.md
… dist/index.html